Moved classes to Model
