+++ src/main/java/seedu/address/MainApp.java
@@ -28,7 +28,7 @@ import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.storage.Storage;
 import seedu.address.storage.StorageManager;
 import seedu.address.storage.UserPrefsStorage;
-import seedu.address.ui.ThemeObserver;
+import seedu.address.model.ThemeObserver;
 import seedu.address.ui.Ui;
 import seedu.address.ui.UiManager;
