+++ src/main/java/seedu/address/model/Model.java
@@ -11,7 +11,6 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.PersonEvent;
 import seedu.address.model.person.PersonStreak;
-import seedu.address.ui.ThemeManager;
 
 /**
  * The API of the Model component.
