+++ src/main/java/seedu/address/ui/UiManager.java
@@ -11,6 +11,7 @@ import seedu.address.MainApp;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.Logic;
+import seedu.address.model.ThemeObserver;
 
 /**
  * The manager of the UI component.
